Move Intel default platform configurations to their own classes, only adding them to default gcc and clang implementations. Relates to Issue: #2909
